requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
openai==1.97.1
pillow==11.3.0
propcache==0.3.2
python-dotenv==1.1.1
pytz==2025.2
redis[hiredis]>=4.5.0
requests==2.32.4
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
